requirements.txt
cookiecutter==1.7.2
pyyaml

# Testing
pytest==5.4.3
pytest-cookies==0.5.1
flake8===3.8.3
black==19.10b0